Moved getExchangeKeys helpers/redux -> helpers/zones + fixed lint errors
