+++ web/src/helpers/redux.js
@@ -1,13 +1,3 @@
-import {
-  flatMap,
-  keys,
-  sortBy,
-  uniq,
-} from 'lodash';
-
-export function getExchangeKeys(zoneHistory) {
-  return sortBy(uniq(flatMap(zoneHistory, d => keys(d.exchange))));
-}
 
 export function getCurrentZoneData(state) {
   const { grid } = state.data;
