+++ web/src/helpers/zones.js
@@ -1,0 +1,10 @@
+import {
+  flatMap,
+  keys,
+  sortBy,
+  uniq,
+} from 'lodash';
+
+export function getExchangeKeys(zoneHistory) {
+  return sortBy(uniq(flatMap(zoneHistory, d => keys(d.exchange))));
+}
